+++ src/api/helpers/requestLogger.ts
@@ -1,14 +1,15 @@
 import type express from 'express';
 
-export default (
+export function requestLogger(
   req: express.Request,
   res: express.Response,
   next: express.NextFunction
-): void => {
+): void {
   if (['/ready', '/healthy'].includes(req.url)) {
     next();
     return;
   }
+
   console.info(`[${req.method}] ${req.url}`);
   next();
-};
+}
