+++ src/api/routes/ready.ts
@@ -1,10 +1,7 @@
 import type express from 'express';
 
-import isReady from 'lib/helpers/isReady';
+import { isReady } from 'lib/helpers/isReady';
 
-export default async function ready(
-  req: express.Request,
-  res: express.Response
-): Promise<void> {
-  res.status((await isReady()) ? 200 : 503).send();
+export function ready(req: express.Request, res: express.Response): void {
+  res.status(isReady() ? 200 : 503).send();
 }
