+++ src/lib/helpers/isReady.ts
@@ -1,5 +1,5 @@
 import renderer from 'lib/rendererSingleton';
 
-export default async function isReady(): Promise<boolean> {
-  return await renderer.ready;
+export function isReady(): boolean {
+  return renderer.ready;
 }
