+++ src/lib/install.ts
@@ -1,4 +1,4 @@
-import getChromiumExecutablePath from 'lib/helpers/getChromiumExecutablePath';
+import { getChromiumExecutablePath } from 'lib/helpers/getChromiumExecutablePath';
 
 (async (): Promise<void> => {
   await Promise.all([getChromiumExecutablePath()]);
